backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt<4.0.0",
    "dnspython>=2.8.0",
    "email-validator>=2.3.0",
    "fastapi>=0.131.0",
    "google-auth>=2.48.0",
    "motor>=3.7.1",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "uvicorn[standard]>=0.41.0",
]
